+++ main-app/src/main/java/txu/report/mainapp/dao/WeeklyReportUserDao.java
@@ -39,7 +39,7 @@ public class WeeklyReportUserDao extends AbstractDao<WeeklyReportEntity> {
 
 
 
-    public List<WeeklyReportEntity> getByDepartmentIdFromTo(Date from, Date to,  Long departmentId) {
+    public List<WeeklyReportEntity> getByDepartmentIdFromTo(Date from, Date to,  Integer departmentId) {
         Query query = getEntityManager().createQuery("SELECT W FROM WeeklyReportEntity AS W WHERE W.uploadedAt >=:from AND W.uploadedAt <=: to AND W.department.id =:departmentId  ORDER BY W.uploadedAt DESC");
         query.setParameter("from", from);
         query.setParameter("to", to);
@@ -48,7 +48,7 @@ public class WeeklyReportUserDao extends AbstractDao<WeeklyReportEntity> {
 
     }
 
-    public WeeklyReportEntity getSingleByDepartmentIdFromTo(Date from, Date to, Long departmentId) {
+    public WeeklyReportEntity getSingleByDepartmentIdFromTo(Date from, Date to, Integer departmentId) {
         Query query = getEntityManager().createQuery("SELECT W FROM WeeklyReportEntity AS W WHERE W.uploadedAt >=:from AND W.uploadedAt <=: to AND W.department.id =:departmentId  ORDER BY W.uploadedAt DESC");
         query.setParameter("from", from);
         query.setParameter("to", to);
